requirements.txt
pandas
numpy
scikit-learn
matplotlib
streamlit
google-generativeai